requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0